==> 02_load_all_data_to_run.R <==

#######################################################################
###
### saving cleaned and formatted datafiles
###
#######################################################################

save(d_surv,file = paste0("datafiles/d_surv.Rdata"))
save(d_fit_hunt_neg,file="datafiles/d_fit_hunt_neg.Rdata")
save(d_fit_hunt_pos,file="datafiles/d_fit_hunt_pos.Rdata")
save(d_fit_sus_cens_posttest,file="datafiles/d_fit_sus_cens_posttest.Rdata")
save(d_fit_sus_cens_postno,file="datafiles/d_fit_sus_cens_postno.Rdata")
save(d_fit_sus_mort_posttest,file="datafiles/d_fit_sus_mort_posttest.Rdata")
save(d_fit_sus_mort_postno,file="datafiles/d_fit_sus_mort_postno.Rdata")
save(d_fit_icap_cens,file="datafiles/d_fit_icap_cens.Rdata")
save(d_fit_icap_mort,file="datafiles/d_fit_icap_mort.Rdata")
save(d_fit_rec_neg_mort,file="datafiles/d_fit_rec_neg_mort.Rdata")
save(d_fit_rec_neg_cens_posttest,file="datafiles/d_fit_rec_neg_cens_posttest.Rdata")
save(d_fit_rec_neg_cens_postno,file="datafiles/d_fit_rec_neg_cens_postno.Rdata")
save(d_fit_rec_pos_mort,file="datafiles/d_fit_rec_pos_mort.Rdata")
save(d_fit_rec_pos_cens,file="datafiles/d_fit_rec_pos_cens.Rdata")
save(d_fit_idead,file="datafiles/d_fit_idead.Rdata")
save(d_fit_endlive,file="datafiles/d_fit_endlive.Rdata")
save(Z_period,file="datafiles/Z_period.Rdata")
save(Z_age,file="datafiles/Z_age.Rdata")
save(num_period,file="datafiles/num_period.Rdata")
save(adj_period,file="datafiles/adj_period.Rdata")
save(weights_period,file="datafiles/weights_period.Rdata")
save(age_lookup_f,file="datafiles/age_lookup_f.Rdata")
save(age_lookup_m,file="datafiles/age_lookup_m.Rdata")
# save(period_effect_survival,file="datafiles/period_effect_survival.Rdata")
save(sus_cens_posttest_age2date,file="datafiles/sus_cens_posttest_age2date.Rdata")
save(sus_cens_postno_age2date,file="datafiles/sus_cens_postno_age2date.Rdata")
save(endlive_age2date,file="datafiles/endlive_age2date.Rdata")
save(sus_mort_posttest_age2date,file="datafiles/sus_mort_posttest_age2date.Rdata")
save(sus_mort_postno_age2date,file="datafiles/sus_mort_postno_age2date.Rdata")
save(icap_cens_age2date,file="datafiles/icap_cens_age2date.Rdata")
save(icap_mort_age2date,file="datafiles/icap_mort_age2date.Rdata")
save(rec_neg_cens_posttest_age2date,file="datafiles/rec_neg_cens_posttest_age2date.Rdata")
save(rec_neg_cens_postno_age2date,file="datafiles/rec_neg_cens_postno_age2date.Rdata")
save(rec_neg_mort_age2date,file="datafiles/rec_neg_mort_age2date.Rdata")
save(rec_pos_mort_age2date,file="datafiles/rec_pos_mort_age2date.Rdata")
save(rec_pos_cens_age2date,file="datafiles/rec_pos_cens_age2date.Rdata")
save(idead_age2date,file="datafiles/idead_age2date.Rdata")
# save(nT_overall_period,file="datafiles/nT_overall_period.Rdata")
# save(nT_overall_period_ext,file="datafiles/nT_overall_period_ext.Rdata")
# save(nT_period_presurv,file="datafiles/nT_period_presurv.Rdata")
# save(nknots_age,file="datafiles/nknots_age.Rdata")
# save(nknots_period,file="datafiles/nknots_period.Rdata")
# save(nT_age_surv,file="datafiles/nT_age_surv.Rdata")
# save(nT_period_surv,file="datafiles/nT_period_surv.Rdata")
# save(n_period_lookup,file="datafiles/n_period_lookup.Rdata")
# save(n_agef,file="datafiles/n_agef.Rdata")
# save(n_agem,file="datafiles/n_agem.Rdata")
# save(n_period,file="datafiles/n_period.Rdata")
save(n_adj_period,file="datafiles/n_adj_period.Rdata")
save(n_adj_sp,file="datafiles/n_adj_sp.Rdata")
save(n_sect,file="datafiles/n_sect.Rdata") 
save(scale,file="datafiles/scale.Rdata") 
save(period_lookup,file="datafiles/period_lookup.Rdata")
save(age_lookup_col_f,file="datafiles/age_lookup_col_f.Rdata")
save(age_lookup_col_m,file="datafiles/age_lookup_col_m.Rdata")
save(sect_hunt_pos,file = "datafiles/sect_hunt_pos.Rdata")
save(sect_hunt_neg,file = "datafiles/sect_hunt_neg.Rdata")

#######################################################################
###
### loading cleaned and formatted datafiles
###
#######################################################################

load("datafiles/d_fit_hunt_neg.Rdata")
load("datafiles/d_fit_hunt_pos.Rdata")
load("datafiles/d_fit_sus_cens_posttest.Rdata")
load("datafiles/d_fit_sus_cens_postno.Rdata")
load("datafiles/d_fit_sus_mort_posttest.Rdata")
load("datafiles/d_fit_sus_mort_postno.Rdata")
load("datafiles/d_fit_icap_cens.Rdata")
load("datafiles/d_fit_icap_mort.Rdata")
load("datafiles/d_fit_rec_neg_mort.Rdata")
load("datafiles/d_fit_rec_neg_cens_posttest.Rdata")
load("datafiles/d_fit_rec_neg_cens_postno.Rdata")
load("datafiles/d_fit_rec_pos_mort.Rdata")
load("datafiles/d_fit_rec_pos_cens.Rdata")
load("datafiles/d_fit_idead.Rdata")
load("datafiles/d_fit_endlive.Rdata")
load("datafiles/Z_period.Rdata")
load("datafiles/Z_age.Rdata")
load("datafiles/num_period.Rdata")
load("datafiles/adj_period.Rdata")
load("datafiles/weights_period.Rdata")
load("datafiles/age_lookup_f.Rdata")
load("datafiles/age_lookup_m.Rdata")
load("datafiles/period_effect_survival.Rdata")
load("datafiles/sus_cens_posttest_age2date.Rdata")
load("datafiles/sus_cens_postno_age2date.Rdata")
load("datafiles/endlive_age2date.Rdata")
load("datafiles/sus_mort_posttest_age2date.Rdata")
load("datafiles/sus_mort_postno_age2date.Rdata")
load("datafiles/icap_cens_age2date.Rdata")
load("datafiles/icap_mort_age2date.Rdata")
load("datafiles/rec_neg_cens_posttest_age2date.Rdata")
load("datafiles/rec_neg_cens_postno_age2date.Rdata")
load("datafiles/rec_neg_mort_age2date.Rdata")
load("datafiles/rec_pos_mort_age2date.Rdata")
load("datafiles/rec_pos_cens_age2date.Rdata")
load("datafiles/idead_age2date.Rdata")
load("datafiles/nT_overall.Rdata")
load("datafiles/nT_period_presurv.Rdata")
load("datafiles/nknots_age.Rdata")
load("datafiles/nknots_period.Rdata")
load("datafiles/nT_age_surv.Rdata")
load("datafiles/nT_period_surv.Rdata")
load("datafiles/n_period_lookup.Rdata")
load("datafiles/n_agef.Rdata")
load("datafiles/n_agem.Rdata")
load("datafiles/n_period.Rdata")
load("datafiles/n_adj_period.Rdata")
load("datafiles/n_adj_sp.Rdata")
load("datafiles/n_sect.Rdata") 
load("datafiles/scale.Rdata") 
load("datafiles/period_lookup.Rdata")
load("datafiles/age_lookup_col_f.Rdata")
load("datafiles/age_lookup_col_m.Rdata")
load("datafiles/sect_hunt_pos.Rdata")
load("datafiles/sect_hunt_neg.Rdata")
